requirements.txt
pytest>=8.2.0
playwright>=1.44.0
pytest-playwright>=0.5.0
requests>=2.31.0
pytest-html>=4.1.1
pytest-check>=2.5.0
allure-pytest>=2.13.5
pytest-xdist>=3.5.0
allure-pytest>=2.15.3
pytest-timeout>=2.3.1
allure-python-commons